refactor: split writer/writer.go into core, pages, and index files
Extract page-writing logic into writer_pages.go and index/footer building
into writer_index.go to reduce the size of the core writer file.
